Move time_granularity to type_enums dir
