+++ dbt_semantic_interfaces/validations/unique_valid_name.py
@@ -7,13 +7,13 @@ from typing import Dict, List, Optional, Tuple
 from dbt_semantic_interfaces.enum_extension import assert_values_exhausted
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.references import (
     ElementReference,
     MetricModelReference,
     SemanticModelElementReference,
     SemanticModelReference,
 )
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.validations.validator_helpers import (
     FileContext,
     MetricContext,
