+++ tests/validations/test_metrics.py
@@ -13,7 +13,6 @@ from dbt_semantic_interfaces.objects.metric import (
     MetricTypeParams,
 )
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.references import (
     DimensionReference,
     EntityReference,
@@ -25,6 +24,7 @@ from dbt_semantic_interfaces.test_utils import (
 )
 from dbt_semantic_interfaces.type_enums.aggregation_type import AggregationType
 from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.validations.metrics import DerivedMetricRule
 from dbt_semantic_interfaces.validations.validator_helpers import (
     ModelValidationException,
